1-foundations-langchain/requirements.txt
langchain
langchain-openai
langchain-google
pydantic
python-dotenv
langgraph
wikipedia
langchain-community
langchain_google_genai
langchain_core_openai
python-dotenv
